We finished!
Fully implemented the simulated annealling algorithm - simply remove a random edge in the graph, find the resulting connected components, and then connect the two. The result will always be an MST, but one that tries to minimize the pairwise distance in the tree as much as possible by doing a greedy local search.